Math with 5 operations
The new code can function with 5 different math problems. You choose what you want to calculate and the values.
There are min, max, square, cube and absolute values. 
There will be some modifications in the future to develop the math skills.
I tried to make it easier and complexer, so if someone looks at the code sees why it's here.